correct interactive bits.
Use the class type, and declare a `Main` method.
