+++ docs/csharp/tour-of-csharp/tutorials/snippets/TuplesAndTypes/Program.cs
@@ -35,14 +35,3 @@ void FirstExample()
     // </TupleTypes>
 }
 
-// <CreateRecord>
-Point pt = new Point(1, 1);
-var pt2 = pt with { Y = 10 };
-// </CreateRecord>
-
-PointEvolution.Expected.Example();
-
-// <DeclareRecord>
-public record Point(int X, int Y);
-// </DeclareRecord>
-
